services/audit-trail-service/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0
alembic>=1.13.0
psycopg2-binary>=2.9

# Validation & Config
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

# Background Tasks
apscheduler>=3.10.4

# Security (JWT deferred, keeping for future)
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7

# AI Provider
google-generativeai>=0.8.0

# Utilities
python-dateutil>=2.8
